requirements.txt
streamlit
python-dotenv
pymysql
sentence-transformers
transformers
accelerate
sqlalchemy